1.JDK版本升级到1.6 2.测试代码按maven规范移入test目录
