+++ src/test/java/org/csource/fastdfs/TestClient.java
@@ -6,7 +6,7 @@
  * Please visit the FastDFS Home Page http://www.csource.org/ for more detail.
  **/
 
-package org.csource.fastdfs.test;
+package org.csource.fastdfs;
 
 import org.csource.common.NameValuePair;
 import org.csource.fastdfs.*;
@@ -18,10 +18,10 @@ import java.net.InetSocketAddress;
  * client test
  *
  * @author Happy Fish / YuQing
- * @version Version 1.20
+ * @version Version 1.18
  */
-public class TestAppender {
-  private TestAppender() {
+public class TestClient {
+  private TestClient() {
   }
 
   /**
@@ -70,8 +70,10 @@ public class TestAppender {
       byte[] file_buff;
       NameValuePair[] meta_list;
       String[] results;
-      String appender_filename;
+      String master_filename;
+      String prefix_name;
       String file_ext_name;
+      String generated_slave_filename;
       int errno;
 
       meta_list = new NameValuePair[4];
@@ -96,12 +98,12 @@ public class TestAppender {
       }
 
       startTime = System.currentTimeMillis();
-      results = client.upload_appender_file(file_buff, "txt", meta_list);
-      System.out.println("upload_appender_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
+      results = client.upload_file(file_buff, "txt", meta_list);
+      System.out.println("upload_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
 
   		/*
   		group_name = "";
-  		results = client.upload_appender_file(group_name, file_buff, "txt", meta_list);
+  		results = client.upload_file(group_name, file_buff, "txt", meta_list);
   		*/
       if (results == null) {
         System.err.println("upload file fail, error code: " + client.getErrorCode());
@@ -157,15 +159,21 @@ public class TestAppender {
         }
 
         file_buff = "this is a slave buff".getBytes(ClientGlobal.g_charset);
-        appender_filename = remote_filename;
+        master_filename = remote_filename;
+        prefix_name = "-part1";
         file_ext_name = "txt";
         startTime = System.currentTimeMillis();
-        errno = client.append_file(group_name, appender_filename, file_buff);
-        System.out.println("append_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
-        if (errno == 0) {
-          System.err.println(client.get_file_info(group_name, appender_filename));
-        } else {
-          System.err.println("append file fail, error no: " + errno);
+        results = client.upload_file(group_name, master_filename, prefix_name, file_buff, file_ext_name, meta_list);
+        System.out.println("upload_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
+        if (results != null) {
+          System.err.println("slave file group_name: " + results[0] + ", remote_filename: " + results[1]);
+
+          generated_slave_filename = ProtoCommon.genSlaveFilename(master_filename, prefix_name, file_ext_name);
+          if (!generated_slave_filename.equals(results[1])) {
+            System.err.println("generated slave file: " + generated_slave_filename + "\n != returned slave file: " + results[1]);
+          }
+
+          System.err.println(client.get_file_info(results[0], results[1]));
         }
 
         startTime = System.currentTimeMillis();
@@ -178,7 +186,7 @@ public class TestAppender {
         }
       }
 
-      results = client.upload_appender_file(local_filename, null, meta_list);
+      results = client.upload_file(local_filename, null, meta_list);
       if (results != null) {
         String file_id;
         int ts;
@@ -220,15 +228,21 @@ public class TestAppender {
           System.err.println("Download file fail, error no: " + errno);
         }
 
-        appender_filename = remote_filename;
+        master_filename = remote_filename;
+        prefix_name = "-part2";
         file_ext_name = null;
         startTime = System.currentTimeMillis();
-        errno = client.append_file(group_name, appender_filename, local_filename);
-        System.out.println("append_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
-        if (errno == 0) {
-          System.err.println(client.get_file_info(group_name, appender_filename));
-        } else {
-          System.err.println("append file fail, error no: " + errno);
+        results = client.upload_file(group_name, master_filename, prefix_name, local_filename, null, meta_list);
+        System.out.println("upload_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
+        if (results != null) {
+          System.err.println("slave file group_name: " + results[0] + ", remote_filename: " + results[1]);
+
+          generated_slave_filename = ProtoCommon.genSlaveFilename(master_filename, prefix_name, file_ext_name);
+          if (!generated_slave_filename.equals(results[1])) {
+            System.err.println("generated slave file: " + generated_slave_filename + "\n != returned slave file: " + results[1]);
+          }
+
+          System.err.println(client.get_file_info(results[0], results[1]));
         }
       }
 
@@ -241,7 +255,7 @@ public class TestAppender {
         file_ext_name = null;
       }
 
-      results = client.upload_appender_file(null, f.length(),
+      results = client.upload_file(null, f.length(),
         new UploadLocalFileSender(local_filename), file_ext_name, meta_list);
       if (results != null) {
         group_name = results[0];
@@ -250,32 +264,20 @@ public class TestAppender {
         System.out.println("group name: " + group_name + ", remote filename: " + remote_filename);
         System.out.println(client.get_file_info(group_name, remote_filename));
 
-        appender_filename = remote_filename;
+        master_filename = remote_filename;
+        prefix_name = "-part3";
         startTime = System.currentTimeMillis();
-        errno = client.append_file(group_name, appender_filename, f.length(), new UploadLocalFileSender(local_filename));
-        System.out.println("append_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
-        if (errno == 0) {
-          System.err.println(client.get_file_info(group_name, appender_filename));
-        } else {
-          System.err.println("append file fail, error no: " + errno);
-        }
+        results = client.upload_file(group_name, master_filename, prefix_name, f.length(), new UploadLocalFileSender(local_filename), file_ext_name, meta_list);
+        System.out.println("upload_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
+        if (results != null) {
+          System.err.println("slave file group_name: " + results[0] + ", remote_filename: " + results[1]);
 
-        startTime = System.currentTimeMillis();
-        errno = client.modify_file(group_name, appender_filename, 0, f.length(), new UploadLocalFileSender(local_filename));
-        System.out.println("modify_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
-        if (errno == 0) {
-          System.err.println(client.get_file_info(group_name, appender_filename));
-        } else {
-          System.err.println("modify file fail, error no: " + errno);
-        }
+          generated_slave_filename = ProtoCommon.genSlaveFilename(master_filename, prefix_name, file_ext_name);
+          if (!generated_slave_filename.equals(results[1])) {
+            System.err.println("generated slave file: " + generated_slave_filename + "\n != returned slave file: " + results[1]);
+          }
 
-        startTime = System.currentTimeMillis();
-        errno = client.truncate_file(group_name, appender_filename);
-        System.out.println("truncate_file time used: " + (System.currentTimeMillis() - startTime) + " ms");
-        if (errno == 0) {
-          System.err.println(client.get_file_info(group_name, appender_filename));
-        } else {
-          System.err.println("truncate file fail, error no: " + errno);
+          System.err.println(client.get_file_info(results[0], results[1]));
         }
       } else {
         System.err.println("Upload file fail, error no: " + errno);
